Allow arguments for assembunny instructions to have arbitrary lengths
The puzzle input for 2016 day 23 contains a copy instructions with a
value of `-16`. The previous method of parsing these instructions
checked the length of the entire instruction in order to determine
what slices to pull out and parse. Now, the entire instruction string
the split on whitespace chars and collected into a vector of string
slices; the number of string slices collected is used to determine the
instruction's type.